Moviedetails dev

+++ database/migrations/2020_12_04_223719_create_producers_table.php
@@ -22,7 +22,7 @@ class CreateProducersTable extends Migration
             $table->string('country')->nullable();
             $table->string('language')->nullable();
             $table->integer('share')->nullable();
-            $table->integer('budget')->nullable();
+            $table->bigInteger('budget')->nullable();
             $table->timestamps();
         });
     }
