+++ database/migrations/2021_01_14_141918_create_distributors_table.php
@@ -19,8 +19,8 @@ class CreateDistributorsTable extends Migration
             $table->string('role')->nullable();
             $table->unsignedBigInteger('country_id');
             $table->date('forecast_release_date');
-            $table->integer('forecast_grant')->nullable();
-            $table->integer('pa_costs')->nullable();
+            $table->bigInteger('forecast_grant')->nullable();
+            $table->bigInteger('pa_costs')->nullable();
 //            $table->unsignedBigInteger('created_by');
 //            $table->unsignedBigInteger('updatedBy')->nullable();
             $table->timestamps();
